split Program class
