+++ PdfScribe/App.config
@@ -36,6 +36,7 @@
     <value>True</value>
    </setting>
    <setting name="StripNoRedistill" serializeAs="String">
+	<!-- This will slow things down because another pass will be made through the postscript file - set it to "False" unless needed -->
     <value>True</value>
    </setting>
   </PdfScribe.Properties.Settings>
