+++ PdfScribe/PdfScribe.csproj
@@ -102,6 +102,7 @@
     <Compile Include="GhostScript64.cs" />
     <Compile Include="NativeMethods.cs" />
     <Compile Include="Program.cs" />
+    <Compile Include="Program_2.cs" />
     <Compile Include="Properties\AssemblyInfo.cs">
       <SubType>Code</SubType>
     </Compile>
